% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertInput.R
\name{convertInput}
\alias{convertInput}
\title{Converts The Format of Input Data Set}
\usage{
convertInput(x)
}
\arguments{
\item{x}{Vector or data-frame column of numeric values. Alternatively, the function selects the \emph{first} column if \code{x} is a multi-column data frame.}
}
\value{
Data frame containing one column ``values'', containing the input data set .
}
\description{
\code{convertInput} accepts either a numeric vector or data frame holding a data set, and returns a single-column data frame. The first column in an input data frame is assumed to be the relevant data set. After sorting the entire data set, the function returns it in one data-frame column named ``values''.
}
\examples{
# Create a numeric dataSet in vector form.
dataSet = rweibull(n = 300, shape = 3.5, scale = 100)

# Convert the data set vector to a single-column data frame.
df = convertInput(dataSet)

}
