% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWeibullParams.R
\name{findWeibullParams}
\alias{findWeibullParams}
\title{Find Weibull Distribution Parameters From a Data Set}
\usage{
findWeibullParams(x)
}
\arguments{
\item{x}{Numeric vector or data-frame column. Where the data frame is multi-column, the function selects the \emph{first} column as data set \code{x}. Another \code{distFit} function \code{\link{convertInput}}, performs this data wrangling.}
}
\value{
A list holding the two parameters of the Weibull distribution:
  \item{shape}{Shape parameter}
  \item{scale}{Scale parameter}
}
\description{
\code{findWeibullParams} uses linear regression, or the ``graphical method'', to estimate Weibull distribution parameters from a data set. Regression equations taken from Wais (2017).
}
\examples{
dataSet = rweibull(n = 300, shape = 3.5, scale = 100)
params = findWeibullParams(dataSet)

params$shape       # example answer 3.752336
params$scale       # example answer 101.3405

}
\references{
Wais, P. (2017) ‘Two and three-parameter Weibull distribution in available wind power analysis’, Renewable Energy, 103, pp. 15–29. doi: 10.1016/J.RENENE.2016.10.041.
}
