% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNormalWeibull.R
\name{fitNormalWeibull}
\alias{fitNormalWeibull}
\title{Evaluate Data-set Fits to Normal and Weibull Distributions}
\usage{
fitNormalWeibull(x)
}
\arguments{
\item{x}{Numeric vector or data-frame column. Where the data frame is multi-column, the function selects the \emph{first} column as data set \code{x}. Another \code{distFit} function \code{\link{convertInput}}, performs this data wrangling.}
}
\value{
A list of eight metrics to evaluate the data-set fit against the normal and Weibull distributions:
  \item{normalADstat}{AD test statistic of data set against a normal distribution.}
  \item{normalADp}{AD test p-value of data set against a normal distribution.}
  \item{normalCVRMSE}{CV(RMSE) as \%, goodness-of-fit index with a normal distribution.}
  \item{normalNMBE}{NMBE as \%, goodness-of-fit index with a normal distribution.}
  \item{weibullADstat}{AD test statistic of data set against a Weibull distribution.}
  \item{weibullADp}{AD test p-value of data set against a Weibull distribution.}
  \item{weibullCVRMSE}{CV(RMSE) as \%, goodness-of-fit index with a Weibull distribution.}
  \item{weibullNMBE}{NMBE as \%, goodness-of-fit index with a Weibull distribution.}
}
\description{
\code{fitNormalWeibull} calculates metrics to evaluate if the data set arose from a candidate distribution. An Anderson-Darling (AD) test compares data set \code{x} with samples values from a specified distribution. When executed against a sample from a closer distribution, AD tests produce smaller test statistics and larger p-values (Keller 2011). The p-value is the probability of an \emph{incorrectly} rejecting the null hypothesis of the same distibution, based on data set \code{x}.

The two AD test results of test statistic and p-value, are accompanied by two goodness-of-fit (GOF) indices recommended by Reddy (2007). The GOF indices are error values expressed as percentages over the entire data set. They are coefficient of variation of the root mean square error CV(RMSE) and normalised mean bias error (NMBE). Bringing these metrics together, the returned list comprises eight values --- four against each of the normal and Weibull distributions.
}
\examples{
\dontrun{Evaluate distributions' fits where data-set shape parameter is 3.0.}
dataSetShape3 = rweibull(n = 300, shape = 3, scale = 1000)
fitMetrics = fitNormalWeibull(dataSetShape3)

\dontrun{The Weibull distribution is superior in all four metrics.}
fitMetrics$normalADstat     # example answer 0.432
fitMetrics$normalADp        # example answer 0.817
fitMetrics$normalCVRMSE       # example answer 2.98
fitMetrics$normalNMBE         # example answer -0.304

fitMetrics$weibullADstat    # example answer 0.3
fitMetrics$weibullADp       # example answer 0.9391
fitMetrics$weibullCVRMSE    # example answer 2.45
fitMetrics$weibullNMBE      # example answer 0.014


\dontrun{Evaluate distributions' fits where data-set shape parameter is 3.6.}
dataSetShape3dot6 = rweibull(n = 300, shape = 3.6, scale = 1000)
fitMetrics = fitNormalWeibull(dataSetShape3dot6)

\dontrun{Now, the normal distribution is superior in three of the four metrics.}
fitMetrics$normalADstat     # example answer 0.225
fitMetrics$normalADp        # example answer 0.9827
fitMetrics$normalCVRMSE     # example answer 1.75
fitMetrics$normalNMBE       # example answer -0.244

fitMetrics$weibullADstat    # example answer 0.3
fitMetrics$weibullADp       # example answer 0.9391
fitMetrics$weibullCVRMSE    # example answer 2.02
fitMetrics$weibullNMBE      # example answer 0.001

}
\references{
Keller, P. (2011) Six Sigma Demystified. 2nd edn. McGraw-Hill.

Reddy, T. A., Maor, I. and Panjapornpon, C. (2007) ‘Calibrating Detailed Building Energy Simulation Programs with Measured Data - Part I: General Methodology (RP-1051)’, HVAC&R Research, 13(2).
}
