% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histWeibull.R
\name{histWeibull}
\alias{histWeibull}
\title{Compare Data-set Histogram and Density to its Weibull Distribution}
\usage{
histWeibull(
  x,
  interval.number = 3,
  base.size = 30,
  label = "Energy (kWh)",
  maxAxisDensity = NULL
)
}
\arguments{
\item{x}{Numeric vector or data-frame column. Where the data frame is multi-column, the function selects the \emph{first} column as data set \code{x}. Another \code{distFit} function \code{\link{convertInput}}, performs this data wrangling.}

\item{interval.number}{(optional) positive integer specifying the number of vertical-axis labels. Use of \code{pretty()} ensures elegant interval spacing.}

\item{base.size}{(optional) positive integer specifying consistent ggplot base_size, text axis label size and legend size.}

\item{label}{(optional) string to label the vertical axis. Ideally, the label describes data-set values.}

\item{maxAxisDensity}{(optional) numeric value specifying the maximum plot density. Useful to extend the vertical axis \emph{if necessary}.}
}
\description{
\code{histWeibull} plots the data set’s histogram and its smoothed density line. For comparision, the second density line represents the Weibull distribution parameterised from the data set.
}
\examples{
dataSet = rweibull(n = 300, shape = 3.5, scale = 1000)

histWeibull(dataSet, interval.number = 6, label = "Enter axis label here")

}
\seealso{
\code{\link[stats]{dweibull}}, \code{\link[ggplot2]{geom_histogram}}, \code{\link[ggplot2]{geom_density}}
}
