% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQcompare.R
\name{QQcompare}
\alias{QQcompare}
\title{Compare Data-set in a Quantile-Quantile Plot}
\usage{
QQcompare(x, interval.number = 3, base.size = 30, label = "Energy (kWh)")
}
\arguments{
\item{x}{Numeric vector or data-frame column. Where the data frame is multi-column, the function selects the \emph{first} column as data set \code{x}. Another \code{distFit} function \code{\link{convertInput}}, performs this data wrangling.}

\item{interval.number}{(optional) positive integer specifying the number of vertical-axis labels. Use of \code{pretty()} ensures elegant interval spacing.}

\item{base.size}{(optional) positive integer to specifying consistent ggplot base_size, text axis label size and legend size.}

\item{label}{(optional) string to label the vertical axis. Ideally, the label describes data-set values.}
}
\description{
\code{QQcompare} plots the quantile-quantile (QQ) line of theoretical normality, for comparision to the data set Z scores. Z-scores standardise the data-set values within their own distribution.
}
\examples{
dataSet = rweibull(n = 300, shape = 3.5, scale = 100)

QQcompare(dataSet)

}
\references{
How to compute the z-score with R:

\url{https://www.r-bloggers.com/2020/02/how-to-compute-the-z-score-with-r/}
}
\seealso{
\code{\link[ggplot2]{geom_qq_line}}, \code{\link[stats]{qqplot}}
}
